+++ examples/blog-refine-nextui/src/pages/products/index.ts
@@ -1,0 +1,4 @@
+export { ProductCreate } from "./create";
+export { ProductEdit } from "./edit";
+export { ProductList } from "./list";
+export { ProductShow } from "./show";