+++ packages/connect/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @refinedev/connect
 
+## 1.0.4
+
+### Patch Changes
+
+-   [#4951](https://github.com/refinedev/refine/pull/4951) [`04837c62077`](https://github.com/refinedev/refine/commit/04837c6207758a7460cfb7a5aff2a104967e20ea) Thanks [@aliemir](https://github.com/aliemir)! - - Update build configuration for `esbuild` to use the shared plugins.
+    -   Fix the lodash replacement plugin to skip redundant files.
+
+## 1.0.3
+
+### Patch Changes
+
+-   [#4951](https://github.com/refinedev/refine/pull/4951) [`04837c62077`](https://github.com/refinedev/refine/commit/04837c6207758a7460cfb7a5aff2a104967e20ea) Thanks [@aliemir](https://github.com/aliemir)! - - Update build configuration for `esbuild` to use the shared plugins.
+    -   Fix the lodash replacement plugin to skip redundant files.
+
 ## 1.0.2
 
 ### Patch Changes